+++ tests/test_encoder.py
@@ -13,12 +13,14 @@ Tests against the encoder module.
 :license:      BSD License
 """
 from __future__ import absolute_import, unicode_literals
-import os
-import io
 from nose.tools import eq_, ok_, raises, nottest
 from segno import consts
 from segno import encoder
 from segno.encoder import Buffer
+try:
+    from .utils import read_matrix
+except (ValueError, SystemError):  # Attempted relative import in non-package
+    from utils import read_matrix
 
 
 def bits(s):
@@ -27,34 +29,6 @@ def bits(s):
 
 def check_prepare_data(expected, data, mode, encoding):
     eq_(expected, tuple(encoder.prepare_data(data, mode, encoding)))
-
-
-def read_matrix(name):
-    """\
-    Helper function to read a matrix from /ref_matrix. The file extension .txt
-    is added automatically.
-
-    :return: A tuple of bytearrays
-    """
-    matrix = []
-    with io.open(os.path.join(os.path.dirname(__file__), 'ref_matrix/{0}.txt'.format(name)), 'rt') as f:
-        for row in f:
-            matrix.append(bytearray([int(i) for i in row if i != '\n']))
-    return tuple(matrix)
-
-
-def read_ref_matrix(name):
-    """\
-    Helper function to read a matrix from /ref. The file extension .txt
-    is added automatically.
-
-    :return: A tuple of bytearrays
-    """
-    matrix = []
-    with io.open(os.path.join(os.path.dirname(__file__), 'ref/{0}.txt'.format(name)), 'rt') as f:
-        for row in f:
-            matrix.append(bytearray([int(i) for i in row if i != '\n']))
-    return tuple(matrix)
 
 
 def test_version_as_str():
@@ -1087,8 +1061,7 @@ def test_figure22_mask0():
     # Figure 22 - Mask 0
     version = consts.VERSION_M4
     matrix = _make_figure22_matrix()
-    matrix, mask = encoder.find_best_mask(matrix, version, 'm', True,
-                                          proposed_mask=0)
+    matrix, mask = encoder.find_best_mask(matrix, version, True, proposed_mask=0)
     # Format info = dark modules
     for i in range(9):
         matrix[8][i] = 0x1
@@ -1103,8 +1076,7 @@ def test_figure22_mask1():
     # Figure 22 - Mask 1
     version = consts.VERSION_M4
     matrix = _make_figure22_matrix()
-    matrix, mask = encoder.find_best_mask(matrix, version, 'm', True,
-                                          proposed_mask=1)
+    matrix, mask = encoder.find_best_mask(matrix, version, True, proposed_mask=1)
     # Format info = dark modules
     for i in range(9):
         matrix[8][i] = 0x1
@@ -1119,8 +1091,7 @@ def test_figure22_mask2():
     # Figure 22 - Mask 2
     version = consts.VERSION_M4
     matrix = _make_figure22_matrix()
-    matrix, mask = encoder.find_best_mask(matrix, version, 'm', True,
-                                          proposed_mask=2)
+    matrix, mask = encoder.find_best_mask(matrix, version, True, proposed_mask=2)
     # Format info = dark modules
     for i in range(9):
         matrix[8][i] = 0x1
@@ -1135,8 +1106,7 @@ def test_figure22_mask3():
     # Figure 22 - Mask 3
     version = consts.VERSION_M4
     matrix = _make_figure22_matrix()
-    matrix, mask = encoder.find_best_mask(matrix, version, 'm', True,
-                                          proposed_mask=3)
+    matrix, mask = encoder.find_best_mask(matrix, version, True, proposed_mask=3)
     # Format info = dark modules
     for i in range(9):
         matrix[8][i] = 0x1
@@ -1150,7 +1120,7 @@ def test_fig23_best_mask():
     # ISO/IEC 18004:2015(E) - 7.8.2 Data mask patterns
     # Figure 23
     matrix = read_matrix('fig-23-unmasked')
-    masked_matrix, mask = encoder.find_best_mask(matrix, 1, error=None, is_micro=False)
+    masked_matrix, mask = encoder.find_best_mask(matrix, 1, is_micro=False)
     eq_(0, mask)
     ref_matrix = read_matrix('fig-23-mask-0')
     eq_(ref_matrix, masked_matrix)
@@ -1167,7 +1137,6 @@ def test_format_info_figure26():
     ref_matrix = read_matrix('fig-26')
     eq_(len(ref_matrix), len(matrix))
     eq_(ref_matrix, matrix)
-
 
 
 if __name__ == '__main__':