+++ tests/test_eps.py
@@ -24,7 +24,7 @@ def test_illegal_color_float():
     color = (.1, 1.1, .1)
     qr = segno.make_qr('test')
     out = io.StringIO()
-    qr.eps(out, color=color)
+    qr.save(out, kind='eps', color=color)
 
 
 @raises(ValueError)
@@ -32,7 +32,7 @@ def test_illegal_color_float2():
     color = (-.1, 1.0, .1)
     qr = segno.make_qr('test')
     out = io.StringIO()
-    qr.eps(out, color=color)
+    qr.save(out, kind='eps', color=color)
 
 
 @raises(ValueError)
@@ -40,7 +40,7 @@ def test_illegal_color_int():
     color = (255, 255, 256)
     qr = segno.make_qr('test')
     out = io.StringIO()
-    qr.eps(out, color=color)
+    qr.save(out, kind='eps', color=color)
 
 
 @raises(ValueError)
@@ -48,20 +48,20 @@ def test_illegal_color_int2():
     color = (-1, 1, 1)
     qr = segno.make_qr('test')
     out = io.StringIO()
-    qr.eps(out, color=color)
+    qr.save(out, kind='eps', color=color)
 
 
 def test_default_color():
     qr = segno.make_qr('test')
     out = io.StringIO()
-    qr.eps(out)
+    qr.save(out, kind='eps')
     ok_('setrgbcolor' not in out.getvalue())
 
 
 def test_color():
     qr = segno.make_qr('test')
     out = io.StringIO()
-    qr.eps(out, color='#195805')
+    qr.save(out, kind='eps', color='#195805')
     ok_('setrgbcolor' in out.getvalue())
 
 
@@ -69,7 +69,7 @@ def test_color_omit_black():
     qr = segno.make_qr('test')
     out = io.StringIO()
     # Black does not need setrgbcolor since it is the default stroke color
-    qr.eps(out, color='#000')
+    qr.save(out, kind='eps', color='#000')
     ok_('setrgbcolor' not in out.getvalue())
 
 
@@ -77,7 +77,7 @@ def test_color_omit_black2():
     qr = segno.make_qr('test')
     out = io.StringIO()
     # Black does not need setrgbcolor since it is the default stroke color
-    qr.eps(out, color='Black')
+    qr.save(out, kind='eps', color='Black')
     ok_('setrgbcolor' not in out.getvalue())
 
 
@@ -85,14 +85,14 @@ def test_color_omit_black3():
     qr = segno.make_qr('test')
     out = io.StringIO()
     # Black does not need setrgbcolor since it is the default stroke color
-    qr.eps(out, color=(0, 0, 0))
+    qr.save(out, kind='eps', color=(0, 0, 0))
     ok_('setrgbcolor' not in out.getvalue())
 
 
 def test_background():
     qr = segno.make_qr('test')
     out = io.StringIO()
-    qr.eps(out, background='#EEE')
+    qr.save(out, kind='eps', background='#EEE')
     ok_('setrgbcolor' in out.getvalue())
     ok_('clippath' in out.getvalue())
 
@@ -100,14 +100,14 @@ def test_background():
 def test_default_scale():
     qr = segno.make_qr('test')
     out = io.StringIO()
-    qr.eps(out)
+    qr.save(out, kind='eps')
     ok_('scale' not in out.getvalue())
 
 
 def test_scale():
     qr = segno.make_qr('test')
     out = io.StringIO()
-    qr.eps(out, scale=2)
+    qr.save(out, kind='eps', scale=2)
     ok_('2 2 scale' in out.getvalue())
 
 
@@ -115,7 +115,7 @@ def test_scale_float():
     qr = segno.make_qr('test')
     out = io.StringIO()
     scale = 1.34
-    qr.eps(out, scale=scale)
+    qr.save(out, kind='eps', scale=scale)
     ok_('{0} {0} scale'.format(scale) in out.getvalue())