+++ CHANGES.rst
@@ -1,6 +1,17 @@
 Changes
 =======
 
+0.1.2 -- 2016-08-17
+-------------------
+* Updated docs
+* Backwards incompatible change: Deprecated "eps", "svg", "png", "pdf", and
+  "txt" methods from QRCode. Use QRCode.save.
+  Methods will be removed in 0.1.3
+* Fixed issue #3 (M1 and M3 codes may have undefined areas)
+* Fixed issue #4 (wrong 'error' default value for encoder.encode(),
+  factory function segno.make() wasn't affected)
+
+
 0.1.1 -- 2016-08-14
 -------------------
 * Initial release